backend/requirements.txt
Django>=4.2
djangorestframework
mysqlclient
django-allauth
djangorestframework-simplejwt
celery
redis
python-dotenv
gunicorn
